test: adapt ActivityInfo retry policy test for Cloud
