+++ packages/test/src/test-local-activities.ts
@@ -1,12 +1,16 @@
 import { firstValueFrom, Subject } from 'rxjs';
 import { Context as ActivityContext } from '@temporalio/activity';
 import { ApplicationFailure, defaultPayloadConverter, WorkflowFailedError } from '@temporalio/client';
+import type { RetryPolicy } from '@temporalio/common';
 import { msToNumber } from '@temporalio/common/lib/time';
 import { temporal } from '@temporalio/proto';
 import * as workflow from '@temporalio/workflow';
 import type { LocalActivityOptions } from '@temporalio/workflow';
 import { Worker } from '@temporalio/test-helpers';
 import { helpers, makeTestFunction } from './helpers-integration';
+import { getRetryPolicyFromActivityInfo } from './workflows/local-activities';
+
+export { getRetryPolicyFromActivityInfo };
 
 const test = makeTestFunction({
   workflowsPath: __filename,
@@ -118,6 +122,31 @@ test.serial('isLocal is set correctly', async (t) => {
   await worker.runUntil(async () => {
     t.is(await executeWorkflow(getIsLocal, { args: [true] }), true);
     t.is(await executeWorkflow(getIsLocal, { args: [false] }), false);
+  });
+});
+
+test.serial('retryPolicy is set correctly', async (t) => {
+  const { executeWorkflow, createWorker } = helpers(t);
+  const worker = await createWorker({
+    maxEagerActivityReservationsPerWorkflowTask: 0,
+    activities: {
+      async retryPolicy(): Promise<object | undefined> {
+        return ActivityContext.current().info.retryPolicy;
+      },
+    },
+  });
+
+  const retryPolicy: RetryPolicy = {
+    backoffCoefficient: 1.5,
+    initialInterval: 2.0,
+    maximumAttempts: 3,
+    maximumInterval: 10.0,
+    nonRetryableErrorTypes: ['nonRetryableError'],
+  };
+
+  await worker.runUntil(async () => {
+    t.deepEqual(await executeWorkflow(getRetryPolicyFromActivityInfo, { args: [retryPolicy, true] }), retryPolicy);
+    t.deepEqual(await executeWorkflow(getRetryPolicyFromActivityInfo, { args: [retryPolicy, false] }), retryPolicy);
   });
 });
